Add files_page_specs.dart. Add second item to BasePageSpecs.floatingActionButtonTargetList.
