+++ .idea/workspace.xml
@@ -4,17 +4,17 @@
     <option name="autoReloadType" value="NONE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="398c2f1d-d59b-4632-b499-7ce03f824c85" name="Changes" comment="Rename BaseUI -&gt; BasePage, BaseUIContents -&gt; BasePageSpecs.&#10;&#10;Delete base_ui_view.dart.">
+    <list default="true" id="398c2f1d-d59b-4632-b499-7ce03f824c85" name="Changes" comment="Update comments and ToDo list.">
+      <change afterPath="$PROJECT_DIR$/lib/files/files_page_specs.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/.filecache" beforeDir="false" afterPath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/.filecache" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/app.dill" beforeDir="false" afterPath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/app.dill" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/kernel_snapshot.d" beforeDir="false" afterPath="$PROJECT_DIR$/.dart_tool/flutter_build/8c9aa6caf79de67e2b2245abfd1b8a6f/kernel_snapshot.d" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.flutter-plugins-dependencies" beforeDir="false" afterPath="$PROJECT_DIR$/.flutter-plugins-dependencies" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/base_page/base_page.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/base_page/base_page.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/lib/base_page/base_page_route_map.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/base_page/base_page_route_map.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/base_page/base_page_specs.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/base_page/base_page_specs.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/lib/kar_kam/kar_kam.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/kar_kam/kar_kam.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/main.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/main.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/kar_kam/kar_kam_ui_contents.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/kar_kam/kar_kam_page_specs.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/settings/settings_ui_contents.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/settings/settings_page_specs.dart" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -37,20 +37,20 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "RunOnceActivity.cidr.known.project.marker": "true",
-    "SHARE_PROJECT_CONFIGURATION_FILES": "true",
-    "cidr.known.project.marker": "true",
-    "dart.analysis.tool.window.visible": "false",
-    "io.flutter.reload.alreadyRun": "true",
-    "last_opened_file_path": "/home/damon/StudioProjects/kar_kam/lib/utils",
-    "settings.editor.selected.configurable": "preferences.pluginManager",
-    "show.migrate.to.gradle.popup": "false"
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.cidr.known.project.marker&quot;: &quot;true&quot;,
+    &quot;SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;,
+    &quot;cidr.known.project.marker&quot;: &quot;true&quot;,
+    &quot;dart.analysis.tool.window.visible&quot;: &quot;false&quot;,
+    &quot;io.flutter.reload.alreadyRun&quot;: &quot;true&quot;,
+    &quot;last_opened_file_path&quot;: &quot;/home/damon/StudioProjects/kar_kam/lib/utils&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;preferences.pluginManager&quot;,
+    &quot;show.migrate.to.gradle.popup&quot;: &quot;false&quot;
   }
-}]]></component>
+}</component>
   <component name="RecentsManager">
     <key name="CopyFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/lib/utils" />
@@ -127,7 +127,14 @@
       <option name="project" value="LOCAL" />
       <updated>1688740317327</updated>
     </task>
-    <option name="localTasksCounter" value="9" />
+    <task id="LOCAL-00009" summary="Update comments and ToDo list.">
+      <created>1688741340364</created>
+      <option name="number" value="00009" />
+      <option name="presentableId" value="LOCAL-00009" />
+      <option name="project" value="LOCAL" />
+      <updated>1688741340364</updated>
+    </task>
+    <option name="localTasksCounter" value="10" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -150,7 +157,8 @@
     <MESSAGE value="Add _SettingsUIContents, a new StatefulWidget in settings_ui_contents.dart." />
     <MESSAGE value="Resolve issue associated with Scaffold.geometryOf being triggered by instance of BottomAppBar.&#10;&#10;This commit includes removal of base_ui_view.dart. The functionality that this wiget offered is subsumed within BaseUI.&#10;&#10;This commit also includes an additional FloatingActionButton as a test.&#10;&#10;Need to tidy up" />
     <MESSAGE value="Rename BaseUI -&gt; BasePage, BaseUIContents -&gt; BasePageSpecs.&#10;&#10;Delete base_ui_view.dart." />
-    <option name="LAST_COMMIT_MESSAGE" value="Rename BaseUI -&gt; BasePage, BaseUIContents -&gt; BasePageSpecs.&#10;&#10;Delete base_ui_view.dart." />
+    <MESSAGE value="Update comments and ToDo list." />
+    <option name="LAST_COMMIT_MESSAGE" value="Update comments and ToDo list." />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
